+++ src/tasks/registry.js
@@ -22,6 +22,7 @@ import { task as transformTask } from "./transform.js";
 import { task as explainTask } from "./explain.js";
 import { restockTasks } from "./restock.js";
 import { wordmathTasks } from "./wordmath.js";
+import { convertTasks } from "./convert.js";
 import { datecalcTasks } from "./datecalc.js";
 import { logicgridTasks } from "./logicgrid.js";
 import { tallyTasks } from "./tally.js";
@@ -38,7 +39,7 @@ import { publicTasks } from "./public.js";
 
 export const tasks = [
   healthTask, helloTask, reasonTask, lookupTask, regexTask, chainTask, transformTask, explainTask,
-  ...restockTasks, ...wordmathTasks, ...datecalcTasks, ...logicgridTasks, ...tallyTasks,
+  ...restockTasks, ...wordmathTasks, ...convertTasks, ...datecalcTasks, ...logicgridTasks, ...tallyTasks,
   ...fanoutTasks, ...followTasks, norelevantTask, ...needleTasks, ...needlehopTasks, ...extractTasks, ...dialogueTasks,
   nearmissTask, ...pagedTasks, typedTask,
   ...publicTasks,